pyproject.toml
[project]
name = "orion"
version = "1.0.0"
description = "Offline Multimodal RAG System"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
authors = [
    {name = "HXMAN76"}
]

dependencies = [
    "fastapi>=0.109.0",
    "uvicorn>=0.27.0",
    "python-multipart>=0.0.6",
    "pydantic>=2.5.0",
    "PyMuPDF>=1.23.0",
    "python-docx>=1.1.0",
    "chromadb>=0.4.22",
    "numpy>=1.24.0",
    "ollama>=0.1.6",
    "tqdm>=4.66.0",
    "tiktoken>=0.5.0",
]

[project.optional-dependencies]
ocr = ["paddleocr>=2.7.0", "paddlepaddle>=2.5.0"]
voice = ["openai-whisper>=20231117", "pyannote.audio>=3.1.0", "librosa>=0.10.0", "soundfile>=0.12.0"]
all = ["orion[ocr,voice]"]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
